@@ -2,7 +2,6 @@ import 'package:appbasce/pages/loading.dart';
 import 'package:flutter/material.dart';
 import 'package:appbasce/services/database.dart';
 import 'package:appbasce/classes/miniTB_prediction_class.dart';
-import 'package:appbasce/classes/profile_class.dart';
 import 'package:material_design_icons_flutter/material_design_icons_flutter.dart';
 import 'package:whatsapp_unilink/whatsapp_unilink.dart';
 import 'package:url_launcher/url_launcher.dart';
@@ -14,16 +13,16 @@ class MiniTBStandings extends StatelessWidget {
       stream: DatabaseService().preds,
       builder: (context, snapshot){
         if (snapshot.hasData) {
-          List<MiniTBPred> preds = snapshot.data;
+          List<MiniTBPred> preds = snapshot.data as List<MiniTBPred>;
           String sponsor = sponsors[DateTime.now().weekday-1];
           return Scaffold(
             appBar: AppBar(
               backgroundColor: Colors.blue[400],
-              title: Text('Mini TB - Classifica'),
+              title: const Text('Mini TB - Classifica'),
               centerTitle: true,
               actions: <Widget>[
                 IconButton(
-                  icon: Icon(Icons.home, color: Colors.white),
+                  icon: const Icon(Icons.home, color: Colors.white),
                   onPressed: () {Navigator.popUntil(context, ModalRoute.withName('/'));},
                 ),
               ],
@@ -72,8 +71,8 @@ class MiniTBStandings extends StatelessWidget {
                       padding: const EdgeInsets.fromLTRB(9, 10, 9, 0),
                       child: Card(
                         child: ListTile(
-                          leading: Icon(MdiIcons.whatsapp),
-                          trailing: Icon(Icons.send),
+                          leading: const Icon(MdiIcons.whatsapp),
+                          trailing: const Icon(Icons.send),
                           title: Text(
                               "Manda classifica alla CB",
                               style: TextStyle(
