@@ -5,9 +5,9 @@ class AuthService {
   final FirebaseAuth _auth = FirebaseAuth.instance;
 
   // sign in with email and password
-  Future<UserCredential> signInWithEmailAndPassword() async {
+  Future<UserCredential?> signInWithEmailAndPassword() async {
     try {
-      Future result = _auth.signInWithEmailAndPassword(email: 'torneobasce@gmail.com', password: 'TyLawson36');
+      Future<UserCredential> result = _auth.signInWithEmailAndPassword(email: 'torneobasce@gmail.com', password: 'TyLawson36');
       return result;
     } catch(e){
       print(e.toString());
